[IMP] resource_booking: add freezegun as a python external dependency
[IMP] resource_booking: fix failling BackendCase. test_booking_from_calendar_view

[IMP] resource_booking: fix typo in help parameter for combination_assignment field from resource_booking_type

[IMP] resource_booking: fix failing PortalCase.test_portal_no_bookings

Test was failling cause there was no bookings link on portal home for external users, then no trigger.
Also I had to tweak eslintrc config cause precommit was failling due to ECMA version.

[IMP] resource_booking: No create/unlink call if nothing to do

[FIX] resource_booking: error CalendarQuickCreate title is not a string when scheduling a booking

[FIX] resource_booking: portal responsive booking calendar table going wider and no x scroll

+++ resource_booking/models/resource_booking.py
@@ -432,8 +432,10 @@ class ResourceBooking(models.Model):
                     to_create.append(meeting_vals)
             else:
                 to_delete |= one.meeting_id
-        to_delete.unlink()
-        _self.env["calendar.event"].create(to_create)
+        if to_delete:
+            to_delete.unlink()
+        if to_create:
+            _self.env["calendar.event"].create(to_create)
 
     @api.constrains("combination_id", "meeting_id", "type_id")
     def _check_scheduling(self):
@@ -697,7 +699,7 @@ class ResourceBooking(models.Model):
                     self.duration, False
                 ),
                 default_resource_booking_ids=[(6, 0, self.ids)],
-                default_name=self.name,
+                default_name=self.name or "",
             ),
             "name": _("Schedule booking"),
             "res_model": "calendar.event",
